Run second half of Codon/AA mut matrix ntbk!
All of the matrix JSONs are now ready to visualize!

Note that I ran this code in ipython (using
%run "[notebookname].ipynb") -- apparently this doesn't modify the
notebook, so that's why it looks like the actual .ipynb file is
unchanged.

Also, the notebook errored out after computing the first JSON
(seq2gene2...) -- the reason for this was a small thing in the
second half of the ntbk that didn't acct for 0-coverage positions.
It's fixed now (will commit changes to ntbk in a sec), and it
shouldn't have impacted the results before that point.

next steps: visualize these jsons