Include *all* mutated pos (with ok cov) in the graph
Previously, just mutated positions in pospair2groupcts were included.
This meant that, probs, a lot of isolated positions that weren't linked
to anything were left out of the graph -- this shouldn't impact
any of the .gv plots, i think. however for the sake of describing this
in the paper including all muts makes sense i think