beef up plots of Bacteria CC 1 and CAMP CC 1
Makes this so that these can be used directly in the paper! nice.

also, I updated my imagemagick installation in conda and I think
that changed the styling of all of my graphviz PNGs somehow?
like they use better resolution and a different font??? idk. it looks
nice so i'm not complaining, just confused lmao.