modify OSA filter to completely rm OSA reads
Rather than just conservatively removing individual alignments from
reads until there was no supp alignment overlap left.

The new approach makes sense if we assume that OSAs are indicative
of molecular chimeras in reads; in this case, completely ignoring
these reads is a defensible course of action. There's probs other
ways to look at this -- and I'm sure that there are cases that
can cause OSAs that are *not* the result of sequencing error --
but this approach is probably ok for now.