Fix nonsyn mutation rates in HMGT for - strand genes
i would be happier if this was formally tested but it does seem
to work, using gene 2561 (prev 2562) in bact2 as sanity check

tldr problem was that positions were going R -> L for - strand genes,
which is usually useful but the CP code I added in earlier tonight
assumed L -> R always. Easiest fix was to just compute the position
range ourselves.