Add only_call_if_rare param to naive p-mut calling
This'll make it easier to apply this to lots of different analyses.

Note that I left the default as NOT using this. Let's see, looking
through the notebooks that use the naive calling function using
$ grep -ri "naively_call_mutation" *.ipynb
--

GENERAL ANALYSES THAT SHOULD EXCLUDE RARE MUTS

-CP 1/2/3 plots
-Target/Decoy approach stuff (just, everything in that notebook, I
 think)
-S/N plots
-NNS/NS plots

GENERAL ANALYSES THAT SHOULD INCLUDE RARE MUTS

-Diversity indices
-HM gene tables
-HM intergenic region tables
-Matrices
-Misc text stuff (at least right now, i think)
-Mut Location viz
-Read clustering for g868 in BACT1

So yeah, this choice of default makes sense. phew. And writing this
out gave me an excuse to go through things again :P