Debug alignments to CAMP; know why discontinuity!
TLDR, need to add virtual reads for positions that are covered but
just relatively low coverage. Investigating r/n why exactly the low-
coverage areas get dropped by LJA -- is it due to the 10x k-mer cov
filter, or is it due to just default behavior? will be interesting
to see. (Ideally it's default behavior, right?)