finish running lja with noec on CAMP as sanity check
tldr, yes, the disconnection due to the low-cov region seems to be
an inherent property of lja since it happened even without any
coverage filtering done by the user. the disconnected bubble section
with a bipartite pattern is still there -- i haven't double checked
that the sequences of the corresponding nodes are identical but their
GC contents, lengths, etc seem identical.

this is good -- it justifies that we can fix this by just adding
more virtual reads to cover low-cov areas