p vs # p-muts figs: merge ish
ideally we'd unify the color legend, etc. but i don't have time for
that (and also i have a suspicion it isn't slightly the same since
i dont think the fdr ests for the bottom figure get down to exaaactly
zero).

could make this better, but this is good enough